Various fixes to address compiler warnings: - removed unused variables and functions - added default cases to switch statements - member initialization order in class constructors - lots of signed/unsigned comparisons fixed either by changing a loop iterator from int to size_t, or by selective type casting.

+++ IlmBase/ImathTest/testJacobiEigenSolver.cpp
@@ -85,8 +85,8 @@ void
 verifyOrthonormal (const TM& A, const typename TM::BaseType threshold)
 {
     const TM prod = A * A.transposed();
-    for (int i = 0; i < TM::dimensions(); ++i)
-        for (int j = 0; j < TM::dimensions(); ++j)
+    for (size_t i = 0; i < TM::dimensions(); ++i)
+        for (size_t j = 0; j < TM::dimensions(); ++j)
             if (i == j) 
                 assert (std::abs (prod[i][j] - 1) < threshold);
             else
@@ -100,8 +100,8 @@ computeThreshold(const TM& A)
    typedef typename TM::BaseType T;
    T maxAbsEntry(0);
 
-   for (int i = 0; i < TM::dimensions(); ++i)
-       for (int j = 0; j < TM::dimensions(); ++j)
+   for (size_t i = 0; i < TM::dimensions(); ++i)
+       for (size_t j = 0; j < TM::dimensions(); ++j)
            maxAbsEntry = std::max (maxAbsEntry, std::abs(A[i][j]));
 
    const T eps = std::numeric_limits<T>::epsilon();
@@ -135,8 +135,8 @@ testJacobiEigenSolver(const TM& A)
 
     // Determinant of A and S
     TM MS;
-    for (int i = 0; i < TM::dimensions(); ++i)
-        for (int j = 0; j < TM::dimensions(); ++j)
+    for (size_t i = 0; i < TM::dimensions(); ++i)
+        for (size_t j = 0; j < TM::dimensions(); ++j)
             if(i == j)
                 MS[i][j] = S[i];
             else
@@ -148,8 +148,8 @@ testJacobiEigenSolver(const TM& A)
     // A = V * S * V^T
     TM MA = V * MS * V.transposed();
 
-    for (int i = 0; i < TM::dimensions(); ++i) 
-        for (int j =0; j < TM::dimensions(); ++j) 
+    for (size_t i = 0; i < TM::dimensions(); ++i) 
+        for (size_t j =0; j < TM::dimensions(); ++j) 
             assert(abs(A[i][j]-MA[i][j]) < threshold);
 }
